+++ ark_auth/src/core/csrf.rs
@@ -3,23 +3,25 @@ use crate::driver;
 use chrono::Utc;
 use time::Duration;
 
-/// Create one CSRF key, value pair. Key must be unique.
+/// Create one CSRF key, value pair with time to live in seconds. Key must be unique.
 pub fn create(
     driver: &driver::Driver,
     service: &Service,
     key: &str,
     value: &str,
+    ttl: i64,
 ) -> Result<Csrf, Error> {
-    delete_by_age(driver)?;
+    delete_by_ttl(driver)?;
 
+    let ttl = Utc::now() + Duration::seconds(ttl);
     driver
-        .csrf_create(key, value, &service.id)
+        .csrf_create(key, value, &ttl, &service.id)
         .map_err(Error::Driver)
 }
 
 /// Read one CSRF key, value pair. CSRF key, value pair is deleted after one read.
 pub fn read_by_key(driver: &driver::Driver, key: &str) -> Result<Option<Csrf>, Error> {
-    delete_by_age(driver)?;
+    delete_by_ttl(driver)?;
 
     driver
         .csrf_read_by_key(key)
@@ -32,11 +34,10 @@ pub fn read_by_key(driver: &driver::Driver, key: &str) -> Result<Option<Csrf>, E
         })
 }
 
-/// Delete many CSRF key, value pairs created more than one hour ago.
-fn delete_by_age(driver: &driver::Driver) -> Result<usize, Error> {
-    let previous_hour = Utc::now() - Duration::hours(1);
-
-    match driver.csrf_delete_by_created_at(&previous_hour) {
+/// Delete many CSRF key, value pairs that have expired using.
+fn delete_by_ttl(driver: &driver::Driver) -> Result<usize, Error> {
+    let now = Utc::now();
+    match driver.csrf_delete_by_ttl(&now) {
         Ok(count) => Ok(count),
         Err(err) => {
             warn!("{}", Error::Driver(err));